call client.close() after use
