+++ corehq/apps/change_feed/management/commands/create_kafka_topics.py
@@ -13,11 +13,11 @@ class Command(BaseCommand):
 
 
 def create_kafka_topics():
-    client = get_simple_kafka_client()
-    for topic in topics.ALL:
-        if client.has_metadata_for_topic(topic):
-            status = "already exists"
-        else:
-            client.ensure_topic_exists(topic, timeout=10)
-            status = "created"
-        print("topic {}: {}".format(status, topic))
+    with get_simple_kafka_client() as client:
+        for topic in topics.ALL:
+            if client.has_metadata_for_topic(topic):
+                status = "already exists"
+            else:
+                client.ensure_topic_exists(topic, timeout=10)
+                status = "created"
+            print("topic {}: {}".format(status, topic))
