+++ corehq/blobs/tests/test_util.py
@@ -1,20 +1,11 @@
-from __future__ import unicode_literals
 from __future__ import absolute_import
-from tempfile import NamedTemporaryFile
+from __future__ import unicode_literals
+
 from unittest import TestCase
 
-import corehq.blobs.util as mod
 from six.moves import range
 
-
-class TestClosingContextProxy(TestCase):
-
-    def test_proxy_is_iterable(self):
-        with NamedTemporaryFile(mode='w+') as tmp:
-            tmp.write("line 1\nline 2\n")
-            tmp.seek(0)
-            proxy = mod.ClosingContextProxy(tmp)
-            self.assertEqual(list(proxy), ["line 1\n", "line 2\n"])
+import corehq.blobs.util as mod
 
 
 class TestRandomUrlId(TestCase):
